JavaScript migration
- Migrating Prism and Clipboard -> webpack
- Additional fixes to code highlighting
- Fixes to Smooth Scroll
- Fixes to re-enable jump menu for versions
- Update Algolia
- Replace Magnific with "Baguettebox"
