+++ assets/scss/app.scss
@@ -1,10 +1,13 @@
+@import './overides/overides';
+
 @import '~bulma';
 @import '~magnific-popup/dist/magnific-popup.css';
 @import '~prismjs/themes/prism.css';
+@import '~docsearch.js/dist/cdn/docsearch.css';
+@import '~baguettebox.js/src/baguetteBox.scss';
 @import '~sass-spacing/mixins.scss';
-//@import 'magnific-popup';
 
-@import './overides/overides';
 @import './elements/elements';
 @import './layout/layouts';
 @import './modules/modules';
+
