+++ assets/scss/elements/_forms.scss
@@ -5,3 +5,7 @@
         transition: all ease-in-out $speed;
     }
 }
+
+.algolia-autocomplete {
+    width: 100%;
+}
