+++ assets/scss/elements/elements.scss
@@ -8,3 +8,4 @@
 @import "_footer.scss";
 @import "_community-notice.scss";
 @import "_notifications.scss";
+@import "_prism.scss";
