+++ templates/_master.twig
@@ -120,18 +120,10 @@
                 <img src="{{ asset('images/bolt.png') }}" width='170' alt="BoltCM" />
             </div>
         </div>
-        <div class="footer-bottom">&copy; Copyright 2017 Bolt CMS</div>
+        <div class="footer-bottom">&copy; Copyright 2018 Bolt CMS</div>
     </footer>
 
     <script src="{{ asset('assets/app.js') }}"></script>
-    <script type="text/javascript">
-    docsearch({
-        apiKey: '1b46aac18cdcead6090088f1994f7ea8',
-        indexName: 'bolt',
-        inputSelector: '#algolia-search',
-        algoliaOptions: { 'facetFilters': [ "version:{{ version }}"] }
-    });
-    </script>
 
 {% block extrafooter %}
 {% endblock extrafooter %}
